Refactor forest and fitting
