--- bark/baselines/__init__.py
@@ -1,1 +1,0 @@
-from .rbf import RBFGP
