--- bark/fitting/__init__.py
@@ -1,4 +1,0 @@
-from .bart.bart import BART as BARK
-from .bart.data import BARKData
-from .bart.params import BARTTrainParams as BARKTrainParams
-from .lgbm_fitting import fit_lgbm_forest, lgbm_to_bark_forest
