--- bark/optimizer/__init__.py
@@ -1,3 +1,0 @@
-from .gbm_model import GbmModel
-from .opt_model import build_opt_model
-from .proposals import propose
