+++ examples/bayes_opt/bofire_bark_study.py
@@ -6,10 +6,10 @@ from bofire.data_models.features.api import CategoricalInput
 from bofire.data_models.strategies.api import RandomStrategy
 
 from bark.fitting.bark_sampler import BARKTrainParams, run_bark_sampler
-from bark.forest_numba import create_empty_forest
-from bark.optimizer import propose
+from bark.forest import create_empty_forest
 from bark.optimizer.opt_core import get_opt_core_from_domain
 from bark.optimizer.opt_model import build_opt_model_from_forest
+from bark.optimizer.proposals import propose
 
 benchmark = Himmelblau()
 domain = benchmark.domain
