+++ examples/bayes_opt/bofire_lgbm_study.py
@@ -11,11 +11,11 @@ from botorch import fit_gpytorch_mll
 
 # with install_import_hook("alfalfa", "beartype.beartype"):
 from bark.bofire_utils.sampling import sample_projected
-from bark.fitting import fit_lgbm_forest, lgbm_to_bark_forest
-from bark.optimizer import propose
+from bark.fitting.lgbm_fitting import fit_lgbm_forest, lgbm_to_bark_forest
 from bark.optimizer.opt_core import get_opt_core_from_domain
 from bark.optimizer.opt_model import build_opt_model_from_forest
-from bark.tree_kernels import LeafGP
+from bark.optimizer.proposals import propose
+from bark.tree_kernels.tree_gps import LeafGP
 from bark.utils.domain import get_feature_types_array
 
 benchmark = Hartmann()
