+++ tests/optimization/test_optimality.py
@@ -5,7 +5,7 @@ from bofire.data_models.domain.api import Domain
 
 from bark.benchmarks import StyblinskiTang
 from bark.fitting.bark_sampler import BARKTrainParams, run_bark_sampler
-from bark.forest_numba import batched_forest_gram_matrix, create_empty_forest
+from bark.forest import batched_forest_gram_matrix, create_empty_forest
 from bark.optimizer import propose
 from bark.optimizer.opt_core import get_opt_core_from_domain
 from bark.optimizer.opt_model import build_opt_model_from_forest
